# Use `esbuild` to transpile frontend assets. (Take II)