--- a/graylog2-web-interface/package.json
+++ b/graylog2-web-interface/package.json
@@ -144,6 +144,7 @@
     "css-loader": "^6.5.1",
     "enzyme": "3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
+    "esbuild-loader": "^2.20.0",
     "eslint-config-graylog": "file:packages/eslint-config-graylog",
     "estraverse-fb": "^1.3.1",
     "express": "^4.16.4",