--- a/graylog2-web-interface/src/components/event-definitions/event-definition-form/FieldsForm.jsx
+++ b/graylog2-web-interface/src/components/event-definitions/event-definition-form/FieldsForm.jsx
@@ -26,7 +26,7 @@ import FieldForm from './FieldForm';
 import FieldsList from './FieldsList';
 
 // Import built-in Field Value Providers
-import {} from './field-value-providers';
+import './field-value-providers';
 
 import commonStyles from '../common/commonStyles.css';