--- a/graylog2-web-interface/src/components/event-definitions/event-definitions/EventDefinitionDescription.jsx
+++ b/graylog2-web-interface/src/components/event-definitions/event-definitions/EventDefinitionDescription.jsx
@@ -17,7 +17,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import moment from 'moment';
-import {} from 'moment-duration-format';
+import 'moment-duration-format';
 import lodash from 'lodash';
 import styled, { css } from 'styled-components';