--- a/graylog2-web-interface/src/components/event-notifications/event-notification-form/EventNotificationFormContainer.jsx
+++ b/graylog2-web-interface/src/components/event-notifications/event-notification-form/EventNotificationFormContainer.jsx
@@ -27,7 +27,7 @@ import EventNotificationForm from './EventNotificationForm';
 
 // Import built-in Event Notification Types
 
-import {} from '../event-notification-types';
+import '../event-notification-types';
 
 const initialValidation = {
   errors: {},