--- a/graylog2-web-interface/src/components/indices/rotation/TimeBasedRotationStrategySummary.jsx
+++ b/graylog2-web-interface/src/components/indices/rotation/TimeBasedRotationStrategySummary.jsx
@@ -17,7 +17,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import moment from 'moment';
-import {} from 'moment-duration-format';
+import 'moment-duration-format';
 
 class TimeBasedRotationStrategySummary extends React.Component {
   static propTypes = {