--- a/graylog2-web-interface/src/pages/AuthenticationPage.tsx
+++ b/graylog2-web-interface/src/pages/AuthenticationPage.tsx
@@ -16,7 +16,7 @@
  */
 import * as React from 'react';
 import { useEffect } from 'react';
-import {} from 'components/authentication/bindings'; // Bind all authentication plugins
+import 'components/authentication/bindings'; // Bind all authentication plugins
 
 import AuthenticationPageNavigation from 'components/authentication/AuthenticationPageNavigation';
 import DocsHelper from 'util/DocsHelper';