--- a/graylog2-web-interface/src/pages/ShowEventNotificationPage.jsx
+++ b/graylog2-web-interface/src/pages/ShowEventNotificationPage.jsx
@@ -30,7 +30,7 @@ import withParams from 'routing/withParams';
 import EventNotificationDetails from 'components/event-notifications/event-notification-details/EventNotificationDetails';
 import EventNotificationActionLinks from 'components/event-notifications/event-notification-details/EventNotificationActionLinks';
 import { EventNotificationsActions } from 'stores/event-notifications/EventNotificationsStore';
-import {} from 'components/event-notifications/event-notification-types';
+import 'components/event-notifications/event-notification-types';
 import EventsPageNavigation from 'components/events/EventsPageNavigation';
 
 const ShowEventDefinitionPage = ({ params: { notificationId } }) => {