**Note:** This is a reissue of #13886, which contained issues, as explained in #14068. This PR is fixing them, by a) fixing up imports for side-effects and b) using the same, consistent target specification for minification.

## Description
<!--- Describe your changes in detail -->

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

This PR is implementing `esbuild-loader` so we can make use of the improved performance of `esbuild` during bundling.

### Before:
```
yarn build  192.80s user 3.71s system 296% cpu 1:06.38 total
```

### After:
```
yarn build  38.77s user 1.81s system 156% cpu 25.961 total
```

The `target` specification is what `browserslist` generates for the `defaults` setting right now as a start. There is no direct integration for it in `esbuild-loader` yet, but writing one is a simple task for a follow-up PR.

Another follow-up could be removing `babel-loader` etc. if we decide that we do not need it anymore.

/nocl

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Refactoring (non-breaking change)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [ ] I have added tests to cover my changes.